+++ jest-puppeteer.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   launch: {
     headless: true,
+    args: ['--no-sandbox', '--disable-setuid-sandbox']
   },
 }