+++ src/methods/canScroll.js
@@ -11,7 +11,7 @@ export default (area) => {
   let scroll = getCurrentScroll(area)
   if (scroll.x || scroll.y) return true
 
-  if (area instanceof HTMLDocument) {
+  if (area instanceof Document) {
     if (area.body) return !!(area.body.scrollTop = 1)
     else return !!(area.documentElement.scrollTop = 1)
   }