+++ src/methods/getAllParentNodes.js
@@ -1,0 +1,18 @@
+/**
+ * @param {DSElement|DSArea} node
+ * @returns {DSElements}
+ */
+export default (node) => {
+  const traverse = (toWatch, index = 0) => {
+    const parent = toWatch[index]?.parentNode
+    if (parent) {
+      toWatch.push(parent)
+      index++
+      return traverse(toWatch, index)
+    } else {
+      return toWatch
+    }
+  }
+
+  return traverse([node])
+}