+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.5.0
+
+- Add dropZones! Solves #115
+- Add a `selectionThreshold` settings, solves #130
+- Add a `refreshMemoryRate` settings, to increase/decrease performance as necessary
+
 # 2.4.4
 
 - Add support for `pointerEvents` (from @HollowMan6) [#143](https://github.com/ThibaultJanBeyer/DragSelect/pull/143) & [#128](https://github.com/ThibaultJanBeyer/DragSelect/pull/128)